Initial commit
This is the first batch of code for the tool "BRPy" I'm currently working on. It can be used to distribute rendering workloads for the 3D graphics software Blender among multiple machines and aims to be user friendly and flexible with minimal setup required.

BRPy acts as the testing ground for development of a network protocol I'd like to call "Blender Render Protocol" or "BRP" for short. This protocol would ideally form the basis for the communication between clients and servers that facilitate the distribution of Blender render jobs in the future and allow for their interoperability. BRPy is the reference implementation of BRP written in Python.

In its current state, BRPy offers CLI applications for the client and server. These worked reasonably well in my testing and allow for the distribution of simple rendering tasks as long as all resources are contained within the .blend file. Usage with big .blend files over slow connections is mostly a pain at the moment, as the client sends one UPLOAD-request, at least one RENDER-request and one DELETE-request in that order, so running the client again causes a reupload of the file.

The code has been written for and tested with Python 3.11.7 that comes bundled with Blender 4.1.1 from blender.org. The just released Blender 4.2.0 still uses Python 3.11.7 and seems to run fine as well. As Blender 4.2 is an LTS release and also comes with plenty of awesome new stuff, I'll target it for development moving forward.