Client: add argparse, split into subcommands
Make use of argparse for better argument parsing and help messages in the client. Adding new functionality will be much easier from now on.

The UPLOAD, RENDER and DELETE requests are now split into subcommands, making the client more flexible and easier to use. The problem of unnecessary reuploads is also solved.