Add option for render output file format
The client can now specify which file format the rendered frames should have with the '-F' option.

The server renders in the specified format or in the one set by the .blend file if unspecified. Invalid options are ignored by Blender, rendering in the format set by the .blend file instead.

As the client doesn't include the bpy module, it doesn't know which file extension to use, so it's told by the server.