Decouple render requests and image transfers
Requests to render frames and the transfer of rendered images are now asynchronous. A server can now ask the client for another frame to render before transmitting the rendered image, so it can get to work on the next frame right away. The speedup from this is greatest for slow network connections and/or when rendering big image files.

This also lays the foundation for nested server topologies.