Handle rendering in new render script + fixes
In order to have more control over the rendering process, the server now offloads render tasks to a render script which utilizes Blender's "bpy" Python module. A thread of the server spawns its worker Blender instance running the script and communicates with it via a socket afterwards.
This change eliminates the delay of having to create a new Blender process for every frame. The speedup is greatest for scenes that take only a short time to render a single frame.

A race condition that caused the client to break off the connection too early is also fixed.

Truncation of the client's server list for rendering is also removed, as a subset of the servers may not be reachable. Better try to connect to all servers and only send requests to the ones that can be connected to.