Support request forwarding
The server has seen big changes to support the forwarding of requests to other servers. Either a server can define its child nodes, or a child node can send a SERVE request to its parent node(s) and register.
It allows for tree-like cluster/grid topologies, so many nodes can be addressed from a single distribution node. This may be especially useful if the connection to the cluster is slow, so it only has to be used once and the other nodes are served internally via a faster connection.
The current implementation is far from perfect, but it works in its basic form.

Output from Blender has also been suppressed, so one can actually see what's going on.